pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "hescape"
version = "0.0.1"
description = "A large scale benchmark for cross-modal histology and spatial gene expression learning"
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [
  { name = "Rushin Gindra", email = "rushin.gindra@helmholtz-munich.de" },
]
authors = [
  { name = "Rushin Gindra" },
]
requires-python = ">=3.10,<3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "anndata",
  # for debug logging (referenced from the issue template)
  "conch",
  "datasets>=4",
  "drvi-py",
  "einops>=0.8.1",
  "gdown>=5.2.0",
  "huggingface-hub>=0.35.1",
  "hydra-core>=1.3.2",
  "hydra-joblib-launcher>=1.2",
  "hydra-submitit-launcher>=1.2",
  "lightning>=2.5.2",
  "local-attention>=1.11.2",
  "open-clip-torch>=3",
  "peft>=0.16",
  "pyyaml>=6.0.2",
  "requests>=2.32.5",
  "scikit-learn>=1.7.1",
  "scikit-misc>=0.5.1",
  "session-info2",
  "timm>=1.0.19",
  "tqdm>=4.67.1",
  "wandb>=0.21",
]
optional-dependencies.dev = [
  "pre-commit",
  "twine>=4.0.2",
]
optional-dependencies.doc = [
  "docutils>=0.8,!=0.18.*,!=0.19.*",
  "ipykernel",
  "ipython",
  "myst-nb>=1.1",
  "pandas",
  # Until pybtex >0.24.0 releases: https://bitbucket.org/pybtex-devs/pybtex/issues/169/
  "setuptools",
  "sphinx>=8.1",
  "sphinx-autodoc-typehints",
  "sphinx-book-theme>=1",
  "sphinx-copybutton",
  "sphinx-tabs",
  "sphinxcontrib-bibtex>=1",
  "sphinxext-opengraph",
]
optional-dependencies.test = [
  "coverage",
  "pytest",
]
# https://docs.pypi.org/project_metadata/#project-urls
urls.Documentation = "https://hescape.readthedocs.io/"
urls.Homepage = "https://github.com/peng-lab/hescape"
urls.Source = "https://github.com/peng-lab/hescape"

[tool.hatch.envs.default]
installer = "uv"
features = [ "dev" ]

[tool.hatch.envs.docs]
features = [ "doc" ]
scripts.build = "sphinx-build -M html docs docs/_build -W {args}"
scripts.open = "python -m webbrowser -t docs/_build/html/index.html"
scripts.clean = "git clean -fdX -- {args:docs}"

# Test the lowest and highest supported Python versions with normal deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "stable" ]
python = [ "3.10", "3.13" ]

# Test the newest supported Python version also with pre-release deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "pre" ]
python = [ "3.13" ]

[tool.hatch.envs.hatch-test]
features = [ "test" ]

[tool.hatch.envs.hatch-test.overrides]
# If the matrix variable `deps` is set to "pre",
# set the environment variable `UV_PRERELEASE` to "allow".
matrix.deps.env-vars = [
  { key = "UV_PRERELEASE", value = "allow", if = [ "pre" ] },
]

[tool.ruff]
line-length = 120
src = [ "src" ]
extend-include = [ "*.ipynb" ]

format.docstring-code-format = true

lint.select = [
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "D",      # pydocstyle
  "E",      # Error detected by Pycodestyle
  "F",      # Errors detected by Pyflakes
  "I",      # isort
  "RUF100", # Report unused noqa directives
  "TID",    # flake8-tidy-imports
  "UP",     # pyupgrade
  "W",      # Warning detected by Pycodestyle
]
lint.ignore = [
  "B006",   # Do not use mutable data structures for argument defaults
  "B007",   # Loop control variable `i` not used within loop body
  "B008",   # Errors from function calls in argument defaults. These are fine when the result is immutable.
  "B023",   # Function definition does not bind loop variable `metadata`
  "BLE001", # Do not catch blind exception: `Exception`
  "D100",   # Missing docstring in public module
  "D101",   # Missing docstring in public class
  # Custom Added by author
  "D102", # Missing docstring in public method
  "D103", # Missing docstring in public function
  "D104", # Missing docstring in public package
  "D105", # __magic__ methods are often self-explanatory, allow missing docstrings
  "D107", # Missing docstring in __init__
  # Disable one in each pair of mutually incompatible rules
  "D203", # We don’t want a blank line before a class docstring
  "D205", # 1 blank line required between summary line and description

  "D213",  # <> We want docstrings to start immediately after the opening triple quote
  "D400",  # first line should end with a period [Bug: doesn’t work with single-line docstrings]
  "D401",  # First line should be in imperative mood; try rephrasing
  "E402",  # Module level import not at top of file
  "E501",  # line too long -> we accept long comment lines; formatter gets rid of long code lines
  "E731",  # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
  "E741",  # allow I, O, l as variable names -> I is the identity matrix
  "F821",  # Undefined name `warnings`
  "F841",  # Local variable `annotations` is assigned to but never used
  "UP038", # Use `X | Y` in `isinstance` call instead of `(X, Y)`
]
lint.per-file-ignores."*/__init__.py" = [ "F401" ]
lint.per-file-ignores."docs/*" = [ "I" ]
lint.per-file-ignores."tests/*" = [ "D" ]
lint.pydocstyle.convention = "numpy"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
xfail_strict = true
addopts = [
  "--import-mode=importlib", # allow using test files with same name
]

[tool.coverage.run]
source = [ "hescape" ]
omit = [
  "**/test_*.py",
]

[tool.cruft]
skip = [
  "tests",
  "src/**/__init__.py",
  "src/**/basic.py",
  "docs/api.md",
  "docs/changelog.md",
  "docs/references.bib",
  "docs/references.md",
  "docs/notebooks/example.ipynb",
]

[tool.uv.sources]
drvi-py = { git = "https://github.com/theislab/drvi.git", rev = "update-scvi-version" }
conch = { git = "https://github.com/mahmoodlab/CONCH" }
